Exam features

+++ ExamCorrectionBackend.Domain/Entities/Exam.cs
@@ -12,8 +12,8 @@ namespace ExamCorrectionBackend.Domain.Entities
             ExamTasks = new HashSet<ExamTask>();
         }
 
-        public long Id { get; set; }
-        public long CourseId { get; set; }
+        public int Id { get; set; }
+        public int CourseId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public DateTime? Date { get; set; }
