+++ ExamCorrectionBackend.Domain/Entities/StudentSolution.cs
@@ -7,8 +7,8 @@ namespace ExamCorrectionBackend.Domain.Entities
 {
     public partial class StudentSolution
     {
-        public long Id { get; set; }
-        public long TaskId { get; set; }
+        public int Id { get; set; }
+        public int TaskId { get; set; }
         public string StudentId { get; set; }
         public string Answer { get; set; }
